+++ examples/saveLoadV1/load_run.py
@@ -7,7 +7,7 @@ from netpyne import specs, sim
 simConfig = specs.SimConfig()                # object of class SimConfig to store simulation configuration
 
 simConfig.duration = 50                      # Duration of the simulation, in ms
-simConfig.dt = 0.05                           # Internal integration timestep, in ms
+simConfig.dt = 0.01                           # Internal integration timestep, in ms
 simConfig.createNEURONObj = True             # create HOC objects when instantiating network
 simConfig.createPyStruct = True              # create Python structure (simulator-independent) when instantiating network
 
@@ -28,3 +28,4 @@ simConfig.addAnalysis('plotConn', True)                                         
 ###############################################################################
 
 sim.loadSimulateAnalyze(filename='V1.json', simConfig=simConfig)    
+#sim.load(filename='V1.json', simConfig=simConfig)