+++ netpyne/cell.py
@@ -184,7 +184,8 @@ class Cell (object):
                         for iseg,seg in enumerate(sec['hSec']):  # set mech params for each segment
                             if type(mechParamValue) in [list]: 
                                 mechParamValueFinal = mechParamValue[iseg]
-                            seg.__getattribute__(mechName).__setattr__(mechParamName,mechParamValueFinal)
+                            if mechParamValueFinal:  # avoid setting None values
+                                seg.__getattribute__(mechName).__setattr__(mechParamName,mechParamValueFinal)
                             
             # add ions
             if 'ions' in sectParams:
@@ -303,7 +304,7 @@ class Cell (object):
             netcon.delay = conn['delay']
             netcon.threshold = conn['threshold']
             conn['hNetcon'] = netcon
-
+            
             # Add plasticity 
             if conn.get('plast'):
                 self._addConnPlasticity(conn['plast'], self.secs[conn['sec']], netcon, 0)
@@ -500,7 +501,7 @@ class Cell (object):
             delays = [params['delay']] * params['synsPerConn']
 
         # Check if target is point process (artificial cell) with V not in section
-        pointp = self._setConnPointP(params, secLabels, weightIndex)
+        pointp, weightIndex = self._setConnPointP(params, secLabels, weightIndex)
         if pointp == -1: return
 
         # Add synaptic mechanisms
@@ -893,9 +894,9 @@ class Cell (object):
 
         if pointp and params['synsPerConn'] > 1: # only single synapse per connection rule allowed
             if sim.cfg.verbose: print '  Error: Multiple synapses per connection rule not allowed for cells where V is not in section (cell gid=%d) '%(self.gid)
-            return -1
+            return -1, weightIndex
 
-        return pointp
+        return pointp, weightIndex
 
     def _setConnSynMechs (self, params, secLabels):
         synsPerConn = params['synsPerConn']