+++ netpyne/network.py
@@ -199,7 +199,7 @@ class Network (object):
         return strParams
 
     ###############################################################################
-    # Calculate distance between 2 segmetns
+    # Calculate distance between 2 segments
     ###############################################################################
     def fromtodistance(self, origin_segment, to_segment):
         h.distance(0, origin_segment.x, sec=origin_segment.sec)
@@ -609,7 +609,6 @@ class Network (object):
         if isinstance(connParam['weight'], list): connParam['weightFromList'] = list(connParam['weight'])  # if weight is a list, copy to weightFromList
         if isinstance(connParam['delay'], list): connParam['delayFromList'] = list(connParam['delay'])  # if delay is a list, copy to delayFromList
         if isinstance(connParam['loc'], list): connParam['locFromList'] = list(connParam['loc'])  # if delay is a list, copy to locFromList
-
         
         orderedPreGids = sorted(preCellsTags.keys())
         orderedPostGids = sorted(postCellsTags.keys())
@@ -698,7 +697,7 @@ class Network (object):
     ###############################################################################
     ### Modify cell params
     ###############################################################################
-    def modifyCells (self, params):
+    def modifyCells (self, params, updateMasterAllCells=False):
         # Instantiate network connections based on the connectivity rules defined in params
         sim.timing('start', 'modifyCellsTime')
         if sim.rank==0: 
@@ -707,7 +706,7 @@ class Network (object):
         for cell in self.cells:
             cell.modify(params)
 
-        if hasattr(sim.net, 'allCells'): 
+        if updateMasterAllCells:
             sim._gatherCells()  # update allCells
 
         sim.timing('stop', 'modifyCellsTime')
@@ -717,7 +716,7 @@ class Network (object):
     ###############################################################################
     ### Modify synMech params
     ###############################################################################
-    def modifySynMechs (self, params):
+    def modifySynMechs (self, params, updateMasterAllCells=False):
         # Instantiate network connections based on the connectivity rules defined in params
         sim.timing('start', 'modifySynMechsTime')
         if sim.rank==0: 
@@ -726,8 +725,8 @@ class Network (object):
         for cell in self.cells:
             cell.modifySynMechs(params)
 
-        if hasattr(sim.net, 'allCells'): 
-            sim._gatherCells()  # update allCells
+        if updateMasterAllCells:
+             sim._gatherCells()  # update allCells
 
         sim.timing('stop', 'modifySynMechsTime')
         if sim.rank == 0 and sim.cfg.timing: print('  Done; syn mechs modification time = %0.2f s.' % sim.timingData['modifySynMechsTime'])
@@ -737,7 +736,7 @@ class Network (object):
     ###############################################################################
     ### Modify conn params
     ###############################################################################
-    def modifyConns (self, params):
+    def modifyConns (self, params, updateMasterAllCells=False):
         # Instantiate network connections based on the connectivity rules defined in params
         sim.timing('start', 'modifyConnsTime')
         if sim.rank==0: 
@@ -746,7 +745,7 @@ class Network (object):
         for cell in self.cells:
             cell.modifyConns(params)
 
-        if hasattr(sim.net, 'allCells'): 
+        if updateMasterAllCells:
             sim._gatherCells()  # update allCells
 
         sim.timing('stop', 'modifyConnsTime')
@@ -756,7 +755,7 @@ class Network (object):
     ###############################################################################
     ### Modify stim source params
     ###############################################################################
-    def modifyStims (self, params):
+    def modifyStims (self, params, updateMasterAllCells=False):
         # Instantiate network connections based on the connectivity rules defined in params
         sim.timing('start', 'modifyStimsTime')
         if sim.rank==0: 
@@ -765,7 +764,7 @@ class Network (object):
         for cell in self.cells:
             cell.modifyStims(params)
 
-        if hasattr(sim.net, 'allCells'): 
+        if updateMasterAllCells:
             sim._gatherCells()  # update allCells
 
         sim.timing('stop', 'modifyStimsTime')